# common-platform-product-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-product-mgmt
  description: Product Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/products'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/products/{productId}':
    get:
      tags:
        - Product
      summary: Get Product by ID
      description: Retrieve a specific product by its unique identifier.
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    put:
      tags:
        - Product
      summary: Update Product
      description: Update the information of an existing product by its unique identifier.
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    delete:
      tags:
        - Product
      summary: Delete Product
      description: Remove an existing product by its unique identifier.
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
...
  '/api/v1/categories/{categoryId}/subtypes/{subtypeId}':
    get:
      tags:
        - Product Category
      summary: Get Category Subtype by ID
      description: Retrieve a specific product subtype using its unique identifier.
      operationId: getSubtype
      parameters:
        - name: categoryId
          in: path
          description: Unique identifier of the product category
          required: true
          schema:
            type: integer
            format: int64
        - name: subtypeId
          in: path
          description: Unique identifier of the product subtype
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the product subtype
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategorySubtypeDTO'
        '404':
          description: Product subtype not found
    put:
      tags:
        - Product Category
      summary: Update Category Subtype
      description: Update an existing product subtype by its unique identifier.
      operationId: updateSubtype
      parameters:
        - name: categoryId
          in: path
          description: Unique identifier of the product category
          required: true
          schema:
            type: integer
            format: int64
        - name: subtypeId
          in: path
          description: Unique identifier of the product subtype
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategorySubtypeDTO'
        required: true
      responses:
        '200':
          description: Product subtype updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategorySubtypeDTO'
        '404':
          description: Product subtype not found
    delete:
      tags:
        - Product Category
      summary: Delete Category Subtype
      description: Remove an existing product subtype by its unique identifier.
      operationId: deleteSubtype
      parameters:
        - name: categoryId
          in: path
          description: Unique identifier of the product category
          required: true
          schema:
            type: integer
            format: int64
        - name: subtypeId
          in: path
          description: Unique identifier of the product subtype
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product subtype deleted successfully
        '404':
          description: Product subtype not found
...
  /api/v1/products/filter:
    post:
      tags:
        - Product
      summary: List Products
      description: Retrieve a paginated list of products.
      operationId: getAllProducts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved the list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseProductDTO'
        '404':
          description: No products found
  /api/v1/products:
    post:
      tags:
        - Product
      summary: Create Product
      description: Create a new product with its associated details.
      operationId: createProduct
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Invalid product data provided
...
  '/api/v1/categories/{categoryId}/subtypes/filter':
    post:
      tags:
        - Product Category
      summary: List Category Subtypes
      description: >-
        Retrieve a paginated list of product subtypes belonging to a specific
        category.
      operationId: getAllSubtypes
      parameters:
        - name: categoryId
          in: path
          description: Unique identifier of the product category
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved product subtypes
          content:
            application/json:
              schema:
                $ref: >-
                  #/components/schemas/PaginationResponseProductCategorySubtypeDTO
        '404':
          description: No subtypes found for the specified category
  '/api/v1/categories/{categoryId}/subtypes':
    post:
      tags:
        - Product Category
      summary: Create Category Subtype
      description: Create a new product subtype under a specific product category.
      operationId: createSubtype
      parameters:
        - name: categoryId
          in: path
          description: Unique identifier of the product category
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategorySubtypeDTO'
        required: true
      responses:
        '201':
          description: Product subtype created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategorySubtypeDTO'
...
components:
  schemas:
...
    ProductDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productId:
          type: integer
          format: int64
          readOnly: true
        productSubtypeId:
          type: integer
          format: int64
        productType:
          type: string
          enum:
            - FINANCIAL
            - NON_FINANCIAL
        productName:
          type: string
        productCode:
          type: string
        productDescription:
          type: string
        productStatus:
          type: string
          enum:
            - PROPOSED
            - ACTIVE
            - RETIRED
        launchDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    ProductCategorySubtypeDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productSubtypeId:
          type: integer
          format: int64
          readOnly: true
        productCategoryId:
          type: integer
          format: int64
        subtypeName:
          type: string
        subtypeDescription:
          type: string
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
...
    PaginationResponseProductDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProductDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    PaginationResponseProductCategorySubtypeDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProductCategorySubtypeDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...